rename to createBreakpoint
